+++ test/parsimmon.test.js
@@ -289,8 +289,8 @@ suite('parser', function() {
           line: 1,
           column: 1
         },
-        expected: ['nothing']}
-        );
+        expected: ['nothing']
+      });
     });
 
     test('composes with existing parsers', function(){
@@ -312,6 +312,52 @@ suite('parser', function() {
       var parser = seq(string('a'), notChar('b').times(5).map(join), notChar('b').or(string('b'))).map(join);
 
       assert.deepEqual(parser.parse('acccccb'), {status: true, value: 'acccccb'});
+    });
+  });
+
+  suite('.tryParse', function() {
+    test('returns just the value', function() {
+      var x = 4;
+      assert.equal(Parsimmon.of(x).tryParse(''), x);
+    });
+
+    test('returns throws on a bad parse', function() {
+      assert.throws(function() {
+        Parsimmon.digit.tryParse('a');
+      });
+    });
+
+    test('thrown error message is equal to formatError', function() {
+      var input = 'a';
+      var parser = Parsimmon.digit;
+      var result = parser.parse(input);
+      var errMsg = Parsimmon.formatError(input, result);
+      try {
+        parser.tryParse(input);
+      } catch (err) {
+        assert.equal(err.message, errMsg);
+      }
+    });
+
+    test('thrown error contains full result object', function() {
+      var input = 'a';
+      var parser = Parsimmon.digit;
+      var result = parser.parse(input);
+      try {
+        parser.tryParse(input);
+      } catch (err) {
+        assert.deepEqual(err.result, result);
+      }
+    });
+
+    test('thrown error message is equal to formatError', function() {
+      var input = 'a';
+      var parser = Parsimmon.digit;
+      try {
+        parser.tryParse(input);
+      } catch (err) {
+        assert.deepEqual(err.result, parser.parse(input));
+      }
     });
   });